2.3
Syntax Updates:

Now better supports arrays and dictionaries.

New Features:

A new special passage, "Sidebar", which displays it's contents in the middle of the sidebar.
A new special passage, "Author", which displays at the top of the sidebar.
Built in styling for buttons.

Issues Fixed:
Additions and comparison issues due to init overriding them