0.2.5
Syntax Changes:
Temporary variables, marked with _var as opposed to $var and displayed [var]

New Features:
Built in "choice" macro, which, when clicked, replaces the clickable text and other options in the choice with the indented content.
Built in "replace" macro, which, when clicked replaces the clickable text with the indented content.
Built in "rand" macro, which takes two arguments, the min and max, and chooses a random value in that range.

See https://docs.pycube.org/ for more info.

Issues Fixed:
Passage links not being recognized